requirements.txt
langchain-core
langchain-community
langgraph
langchain-groq
python-dotenv
huggingface-hub
transformers
sentence-transformers
google-api-python-client
google-auth<2.0
google-auth-oauthlib
google-auth-httplib2
faiss-cpu
bs4
langchain-huggingface
langchain-pinecone
pinecone
pinecone-client
streamlit